requirements.txt
httpx>=0.27.0
beautifulsoup4>=4.12.0
pydantic>=2.0
pydantic-settings>=2.0
pyyaml>=6.0
lxml>=5.0
anthropic>=0.40.0
tenacity>=8.0
python-dotenv>=1.0
